requirements.txt
anthropic
discord.py
feedparser
openpyxl
python-dotenv
requests
yfinance
